Changed int to double in Locust

+++ SP1Framework/BossStage2Locust.h
@@ -12,7 +12,8 @@ class BossStage2Locust :
     public Enemy
 {
 private:
-    int timeStartedWaiting;
+
+    double timeStartedWaiting;
     int currentState;
 
     int futureX;
